pyproject.toml
[tool.poetry]
name = "eventhub-api"
version = "0.1.0"
description = "Event managment platform"
authors = ["Artem Degtyarev https://t.me/JohnyBlacksad"]
license = "CC BY-NC-SA 4.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.135.1"
uvicorn = "^0.41.0"
motor = "^3.7.1"
pydantic = "^2.12.5"
pydantic-settings = "^2.13.1"
PyJWT = "^2.11.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
email-validator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^1.3.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
httpx = "^0.26.0"
mongomock = "^4.1.2"
allure-pytest = "^2.13.2"
ruff = "^0.1.14"
faker = "^40.11.0"
mongomock-motor = "^0.0.36"
pytest-subtests = "^0.15.0"
mypy = "^1.19.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

markers = [
    "UNITS: Unit tests",
    "INTEGRATION: Integration tests",
    "LOAD: Load tests",
    "DAO: Data Access Object tests",
    "SERVICES: Service layer tests",
    "API: API endpoint tests",
    "USERS: User related tests",
    "EVENTS: Event related tests",
    "AUTH: Authentication tests",
    "ACTIVATION_CODE: Activation code tests",
    "EVENT_REGISTRATION: Event registration tests",
    "CREATE_USER: Testing user creation logic",
    "GET_USER: Testing get user logic",
    "UPDATE_USER: Testing user update logic",
    "DELETE_USER: Testing user deletion logic",
    "UPDATE_ROLE: Testing user role update logic",
    "BAN_USER: Testing user ban logic",
    "GET_USERS: Testing user list and filtering logic",
    "SETUP_INDEXES: Testing database index setup",
    "CREATE_EVENT: Testing event creation logic",
    "GET_EVENT: Testing get event logic",
    "UPDATE_EVENT: Testing update event logic",
    "DELETE_EVENT: Tesing delete event logic",
    "HAS_ACTIVE_EVENT: Testing active status event logic",
    "DELETE_USER_EVENTS: Testing delete all user events",
    "GET_EVENTS: Tesing event list and filtering logic",
    "ADD_REGISTRATION: Testing registration in event logic",
    "REMOVE_REGISTRATION: Testing remove registration in event logic",
    "GET_EVENT_REGISTRATIONS: Testing get registration in event logic",
    "GET_USER_REGISTRATIONS: Testing get registration user in event logic",
    "GET_EXISTING_REGISTRATION: Testing get existing registartion in event logic",
    "DELETE_ALL_REGISTRATIONS_FOR_EVENT: Testing delete all registration in event logic",
    "SET_DELETION_TIME_FOR_EVENT: Tesing set deletion time for event logic",
    "DELETE_REGISTRATION_BY_USER: Testing delete all registration user in event logic",
    "CREATE_CODE: Testsing create update role code logic",
    "GET_CODE: Testing get activation code logic",
    "GET_CODES: Testing get acivation codes list logic",
    "USE_CODE: Teting use activation code logic",
    "DELETE_CODE: Testing delete activation code logic",
    "FILTER_CODES: Testing get filtered codes list logic",
    "HASH_PASSWORD: Testing hash password logic",
    "VERIFY_PASSWORD: Testing verify password logic",
    "CREATE_TOKEN: Testing creating token logic",
    "DECODE_TOKEN: Testing decode token logic",
    "REFRESH_TOKEN: Testing refresg token logic"
]

addopts = [
    "-s",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=xml:./tests/reports/allure-results/coverage.xml",
    "--cov-report=html:./tests/reports/coverage-html",
    "--alluredir=./tests/reports/allure-results",
    "--verbose"
]


[tool.ruff]
select = ["E", "F", "I", "B", "UP", "SIM", "C4"]
line-length = 120

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true